Moved selectors to routinesSlice and added new routine name logic
